pyproject.toml
[build-system]
requires = [
    "setuptools>=68",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "predictapool-backend"
version = "0.1.0"
description = "Predict-A-Pool backend service"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
    "strawberry-graphql>=0.224",
    "passlib>=1.7.4",
    "bcrypt<4.0.1",
    "pydantic[email]",
    "asyncpg",
    "greenlet",
    "python-jose[cryptography]"
]

[tool.uvicorn]
factory=false

[tool.setuptools]
packages = ["app"]